+++ src/app/@modal/(.)signup/page.tsx
@@ -4,36 +4,13 @@ import React, { useState } from "react";
 import Image from "next/image";
 import Button from "@/components/atoms/Button";
 import ConsentCheckbox from "@/components/atoms/CheckBox";
-import Dropdown from "@/components/atoms/Dropdown";
 import Modal from "@/components/atoms/Modal";
 import InputBox from "@/components/molecules/InputBox";
+import DropdownBox from "@/components/molecules/DropdownBox";
 import { validateEmail, validatePassword, validateName } from "@/utils/validation";
 
 function SignupHome() {
-  const createOption = (value: string) => ({ value, label: value });
-
-  const options1 = [createOption("광역시 / 도"), createOption("서울특별시"), createOption("부산광역시")];
-
-  const options2 = [createOption("시 / 군 / 구"), createOption("남구"), createOption("북구")];
-
-  const options3 = [createOption("동 / 면 / 읍"), createOption("장전동"), createOption("부저동")];
-
-  const [selectedOption1, setSelectedOption1] = useState(options1[0].value);
-  const [selectedOption2, setSelectedOption2] = useState(options2[0].value);
-  const [selectedOption3, setSelectedOption3] = useState(options3[0].value);
   const [consentChecked, setConsentChecked] = useState(false); // 동의 체크 상태
-
-  const handleDropdownChange1 = (selectedValue: any) => {
-    setSelectedOption1(selectedValue);
-  };
-
-  const handleDropdownChange2 = (selectedValue: any) => {
-    setSelectedOption2(selectedValue);
-  };
-
-  const handleDropdownChange3 = (selectedValue: any) => {
-    setSelectedOption3(selectedValue);
-  };
 
   const handleButtonClick = () => {
     console.log("회원가입 버튼 클릭");
@@ -77,29 +54,8 @@ function SignupHome() {
           />
         </div>
         {/* 지역 선택 */}
+        <DropdownBox />
 
-        <div className="flex justify-between">
-          <Dropdown
-            options={options1}
-            selectedValue={selectedOption1}
-            onChange={handleDropdownChange1}
-            className="w-[180px]" // Adjust the width as needed
-          />
-
-          <Dropdown
-            options={options2}
-            selectedValue={selectedOption2}
-            onChange={handleDropdownChange2}
-            className="w-[180px]" // Adjust the width as needed
-          />
-
-          <Dropdown
-            options={options3}
-            selectedValue={selectedOption3}
-            onChange={handleDropdownChange3}
-            className="w-[180px]" // Adjust the width as needed
-          />
-        </div>
         <p className="pb-[32px]" />
         {/* 개인 정보 수집 및 이용 동의 체크박스 */}
         <div className="mb-4">
